Move the menus plugin logic to app level
